requirements.txt
streamlit>=1.31.0
google-generativeai>=0.8.0
python-docx>=1.1.0
pdfplumber>=0.11.0
python-pptx>=0.6.23
openpyxl>=3.1.0
